abstraction-layer/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
respx>=0.22.0
pytest-asyncio>=0.24.0
boto3>=1.42.0
stripe>=9.0.0
email-validator>=2.0.0
